refactor: Mudar tipo do atributo entregaId de Entrega de Long para Integer

+++ src/main/java/com/inventario/projeto/model/Entrega.java
@@ -16,7 +16,7 @@ public class Entrega {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long entregaId;
+    private Integer entregaId;
 
     @ManyToOne
     @JoinColumn(name = "movimentaco_id")
